@@ -4,7 +4,7 @@ import { Box } from 'boxible'
 import styled from '@emotion/styled'
 import { cx } from '@emotion/css'
 import { css, keyframes } from '@emotion/react'
-import { useLocation, useHistory, useParams } from 'react-router-dom'
+import { useLocation, useNavigate, useParams } from 'react-router-dom'
 import dayjs from 'dayjs'
 import localizedFormat from 'dayjs/plugin/localizedFormat'
 
@@ -14,5 +14,5 @@ export { field, modelize, hydrateModel, hydrateInstance } from 'modeled-mobx'
 
 export {
     React, ReactDOM, Box, styled, useState, useEffect, useContext, useMemo, useCallback, Component,
-    css, keyframes, cx, useLocation, useHistory, useParams,
+    css, keyframes, cx, useLocation, useNavigate, useParams,
 }
