@@ -1,5 +1,6 @@
 export { Box } from 'boxible'
 export * from './button'
+export * from './styled-buttons'
 export * from './icon'
 export * from './modal'
 export * from './form'
@@ -13,7 +14,6 @@ export * from './loading-animation'
 export * from './ui-states'
 export * from './col'
 export * from './incorrect-user'
-export * from './logout-button'
 export * from './menu'
 export * from './footer'
 export * from './off-canvas'
